Screen/Color: pass uint8_t component values to constructor
These values are 0..255, and the API should reflect that.

+++ src/Gauge/GaugeVario.cpp
@@ -117,7 +117,7 @@ GaugeVario::GaugeVario(ContainerWindow &parent,
   } else {
     colText = Color::BLACK;
     colTextBackgnd = Color::WHITE;
-    colTextGray = Color(~0xa0, ~0xa0, ~0xa0);
+    colTextGray = Color((uint8_t)~0xa0, (uint8_t)~0xa0, (uint8_t)~0xa0);
     hBitmapClimb.load(IDB_CLIMBSMALL);
   }
